Renderizado desde de component, callback
